[MOL-22129][XF] fix accordian e2e test
